Rename VoteStatus to StakeStatus
