+++ src/modules/pos/validator/components/StakeStatus/Status.js
