+++ src/modules/pos/validator/components/StakeStatus/index.js
