+++ src/modules/pos/validator/components/StakeStatus/status.test.js
@@ -6,7 +6,7 @@ import Result from './Status';
 const props = {
   t: (s) => s,
   transactions: { txSignatureError: null, signedTransaction: {} },
-  statusInfo: { locked: 200, unlockable: 100, selfUnvote: undefined },
+  statusInfo: { locked: 200, unlockable: 100, selfUnStake: undefined },
 };
 
 jest.mock('@libs/wcm/hooks/useSession', () => ({
@@ -25,19 +25,19 @@ describe('VotingQueue.Result', () => {
   it('renders properly', () => {
     const wrapper = mountWithRouter(Result, props);
 
-    expect(wrapper).toContainMatchingElement('VoteSuccessfulModal');
+    expect(wrapper).toContainMatchingElement('StakeSuccessfulModal');
   });
 
   it('displays the locked message properly', () => {
     const wrapper = mountWithRouter(Result, { ...props, statusInfo: { locked: 200 } });
-    const element = wrapper.find('VoteSuccessfulModal');
+    const element = wrapper.find('StakeSuccessfulModal');
 
     expect(element.text()).toContain('0.000002 LSK will be locked for voting.');
   });
 
   it('displays the unlocked message properly', () => {
     const wrapper = mountWithRouter(Result, { ...props, statusInfo: { unlockable: 300 } });
-    const element = wrapper.find('VoteSuccessfulModal');
+    const element = wrapper.find('StakeSuccessfulModal');
 
     expect(element.text()).toContain(
       '0.000003 LSK will be available to unlock in {{unlockTime}}h.'
@@ -49,7 +49,7 @@ describe('VotingQueue.Result', () => {
       ...props,
       statusInfo: { locked: 200, unlockable: 300 },
     });
-    const element = wrapper.find('VoteSuccessfulModal');
+    const element = wrapper.find('StakeSuccessfulModal');
 
     expect(element.text()).toContain(
       'You have now locked 0.000002 LSK for voting and may unlock 0.000003 LSK in {{unlockTime}} hours.'
