+++ src/modules/pos/validator/components/StakeStatus/styles.css
